# Further improvement of buildpack URIs support