--- a/buildpackage/config_reader.go
+++ b/buildpackage/config_reader.go
@@ -6,9 +6,9 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/pkg/errors"
 
+	"github.com/buildpacks/pack/internal/buildpack"
 	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/dist"
-	"github.com/buildpacks/pack/internal/paths"
 	"github.com/buildpacks/pack/internal/style"
 )
 
@@ -79,24 +79,11 @@ func (r *ConfigReader) Read(path string) (Config, error) {
 		return packageConfig, err
 	}
 
-	absPath, err := paths.ToAbsolute(packageConfig.Buildpack.URI, configDir)
-	if err != nil {
-		return packageConfig, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(packageConfig.Buildpack.URI))
+	if err := validateURI(packageConfig.Buildpack.URI, configDir); err != nil {
+		return packageConfig, err
 	}
-	packageConfig.Buildpack.URI = absPath
 
-	for i := range packageConfig.Dependencies {
-		uri := packageConfig.Dependencies[i].URI
-		if uri != "" {
-			absPath, err := paths.ToAbsolute(uri, configDir)
-			if err != nil {
-				return packageConfig, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(uri))
-			}
-
-			packageConfig.Dependencies[i].URI = absPath
-		}
-
-		dep := packageConfig.Dependencies[i]
+	for _, dep := range packageConfig.Dependencies {
 		if dep.URI != "" && dep.ImageName != "" {
 			return packageConfig, errors.Errorf(
 				"dependency configured with both %s and %s",
@@ -104,7 +91,26 @@ func (r *ConfigReader) Read(path string) (Config, error) {
 				style.Symbol("image"),
 			)
 		}
+
+		if dep.URI != "" {
+			if err := validateURI(dep.URI, configDir); err != nil {
+				return packageConfig, err
+			}
+		}
 	}
 
 	return packageConfig, nil
+}
+
+func validateURI(uri, relativeBaseDir string) error {
+	locatorType, err := buildpack.GetLocatorType(uri, relativeBaseDir, nil)
+	if err != nil {
+		return err
+	}
+
+	if locatorType == buildpack.InvalidLocator {
+		return errors.Errorf("invalid locator %s", style.Symbol(uri))
+	}
+
+	return nil
 }