--- a/internal/buildpack/parse_name.go
+++ b/internal/buildpack/parse_name.go
@@ -5,14 +5,16 @@ import (
 	"strings"
 )
 
-func ParseLocator(locator string) (id string) {
-	return ParseRegistryLocator(ParseBuilderLocator(ParsePackageLocator(locator)))
-}
-
-// ParseIDLocator parses a buildpack locator of the form <id>@<version> into its ID and version.
+// ParseIDLocator parses a buildpack locator in the following formats into its ID and version.
+//
+//   - <id>[@<version>]
+//   - urn:cnb:builder:<id>[@<version>]
+//   - urn:cnb:registry:<id>[@<version>]
+//   - from=builder:<id>[@<version>] (deprecated)
+//
 // If version is omitted, the version returned will be empty. Any "from=builder:" or "urn:cnb" prefix will be ignored.
 func ParseIDLocator(locator string) (id string, version string) {
-	nakedLocator := ParseLocator(locator)
+	nakedLocator := parseRegistryLocator(parseBuilderLocator(locator))
 
 	parts := strings.Split(nakedLocator, "@")
 	if len(parts) == 2 {
@@ -21,30 +23,38 @@ func ParseIDLocator(locator string) (id string, version string) {
 	return parts[0], ""
 }
 
-func ParseRegistryID(registryID string) (namespace string, name string, version string, err error) {
-	id, version := ParseIDLocator(registryID)
-
-	parts := strings.Split(id, "/")
-	if len(parts) == 2 {
-		return parts[0], parts[1], version, nil
-	}
-	return parts[0], "", version, fmt.Errorf("invalid registry ID: %s", registryID)
-}
-
-func ParseRegistryLocator(locator string) (path string) {
-	return strings.TrimPrefix(locator, fromRegistryPrefix+":")
-}
-
-func ParseBuilderLocator(locator string) (path string) {
-	return strings.TrimPrefix(
-		strings.TrimPrefix(locator, deprecatedFromBuilderPrefix+":"),
-		fromBuilderPrefix+":")
-}
-
-func ParsePackageLocator(locator string) (path string) {
+// ParsePackageLocator parses a locator (in format `[docker://][<host>]/<path>[:<tag>⏐@<digest>]`) in image name (`[<host>]/<path>[:<tag>⏐@<digest>]`)
+func ParsePackageLocator(locator string) (imageName string) {
 	return strings.TrimPrefix(
 		strings.TrimPrefix(
 			strings.TrimPrefix(locator, fromDockerPrefix+"//"),
 			fromDockerPrefix+"/"),
 		fromDockerPrefix)
+}
+
+// ParseRegistryID parses a registry id (ie. `<namespace>/<name>@<version>`) into namespace, name and version components.
+//
+// Supported formats:
+//   - <ns>/<name>[@<version>]
+//   - urn:cnb:registry:<ns>/<name>[@<version>]
+//
+func ParseRegistryID(registryID string) (namespace string, name string, version string, err error) {
+	id, version := ParseIDLocator(registryID)
+
+	parts := strings.Split(id, "/")
+	if len(parts) != 2 {
+		return "", "", "", fmt.Errorf("invalid registry ID: %s", registryID)
+	}
+
+	return parts[0], parts[1], version, nil
+}
+
+func parseRegistryLocator(locator string) (path string) {
+	return strings.TrimPrefix(locator, fromRegistryPrefix+":")
+}
+
+func parseBuilderLocator(locator string) (path string) {
+	return strings.TrimPrefix(
+		strings.TrimPrefix(locator, deprecatedFromBuilderPrefix+":"),
+		fromBuilderPrefix+":")
 }